requirements.txt
fastapi
python-dotenv
uvicorn
pydantic
pandas
scikit-learn==1.7.1
joblib==1.5.1
mlflow==3.1.4
prometheus_client==0.22.1
dvc==3.61.0
